refactor(runtime): split regex_eval.rs into cohesive modules (§7-8)
Split the 973-line regex_eval.rs (over the 500-line limit) into three
cohesive modules, all under 500 lines:

- regex_eval.rs (397)        - registry copy, regex closure/code-assertion
                              eval, named-capture/reduce-time actions
- regex_eval_repeat.rs (323) - quantifier repeat-code + eager code blocks
- regex_eval_class.rs (259)  - char-class matching + find-all

All methods are impl Interpreter, so the split is visibility-only. The
regex_helpers import (CaseFoldIter/matches_named_builtin) moves to the
char-class module that uses it. The fn multiset is byte-identical. No
behavior change.
