+++ src/runtime/regex/mod.rs
@@ -1,5 +1,7 @@
 mod regex_casefold;
 mod regex_eval;
+mod regex_eval_class;
+mod regex_eval_repeat;
 pub(crate) mod regex_helpers;
 mod regex_interpolate;
 mod regex_match_atom;
